pyproject.toml
[tool.poetry]
name = "charity"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
django-mdeditor = "^0.1.20"
djangorestframework = "^3.15.1"
ruff = "^0.3.4"
django-storages = {extras = ["s3"], version = "^1.14.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 90